refactor RicettaDettagliataFragment e CreaRicettaFragment
