+++ app/src/main/java/it/unimib/brewday/ui/gestisci_ricette/CreaRicettaFragment.java
@@ -27,15 +27,13 @@ import java.util.List;
 import it.unimib.brewday.R;
 import it.unimib.brewday.model.Ingrediente;
 import it.unimib.brewday.ui.gestisci_ingredienti.AdapterListViewListaIngredientiDisponibili;
+import it.unimib.brewday.util.GestioneRicette;
 import it.unimib.brewday.util.ListaIngredienti;
 
 public class CreaRicettaFragment extends Fragment {
 
     private  FragmentCreaRicettaBinding fragmentCreaRicettaBinding;
-    int posizionePrecedente = -1;
-    EditText quantitaIngredientePrecedente;
 
-    Ingrediente ingredientePrecedente;
     List<Ingrediente> listaIngredientiRicetta;
 
 
@@ -64,7 +62,7 @@ public class CreaRicettaFragment extends Fragment {
         Button creaRicettaButton = fragmentCreaRicettaBinding.buttonCreaRicetta;
         EditText numeroLitriBirra = fragmentCreaRicettaBinding.editNumberNumeroLitriBirra;
         EditText nomeRicetta = fragmentCreaRicettaBinding.editTextNomeRicetta;
-
+        GestioneRicette gestioneRicette = new GestioneRicette();
 
         ListaIngredienti listaIngredienti = new ListaIngredienti();
         listaIngredientiRicetta = listaIngredienti.getListaIngredienti();
@@ -90,12 +88,12 @@ public class CreaRicettaFragment extends Fragment {
         listViewIngredientiRicetta.setDivider(null);
 
         numeroLitriBirra.setOnFocusChangeListener((v, hasFocus) ->
-               verificaNumeroLitriBirra(numeroLitriBirra, hasFocus)
+                gestioneRicette.verificaNumeroLitriBirra(numeroLitriBirra, hasFocus)
         );
 
         creaRicettaButton.setOnClickListener(v -> {
 
-                    if( controlloCreazione(view, nomeRicetta, numeroLitriBirra)){
+                    if( gestioneRicette.controlloCreazione(view, nomeRicetta, numeroLitriBirra)){
                     int zeroIngredinti = 0;
                     double litriScelti = Double.parseDouble(numeroLitriBirra.getText().toString());
                     List<Double> listaIngredientiPerLitro = new ArrayList<>();
@@ -111,95 +109,11 @@ public class CreaRicettaFragment extends Fragment {
         });
     }
 
-    public Ingrediente verificaIngrediente(Ingrediente ingrediente, EditText quantitaIngrediente) {
-
-        if (quantitaIngrediente.getText().length() == 0) {
-            ingrediente.setQuantitaPosseduta(0);
-            quantitaIngrediente.setText("0");
-        } else {
-            quantitaIngrediente.setText(String.valueOf(Integer.parseInt(quantitaIngrediente.getText().toString())));
-            ingrediente.setQuantitaPosseduta(Integer.parseInt(quantitaIngrediente.getText().toString()));
-        }
-
-        return ingrediente;
-    }
-
-    public int quantitaBottone(int position) {
-        if (position == 0) {
-            return 1;
-        } else {
-            return 10;
-        }
 
 
-    }
-    public void inizializzaPositionePrecedente(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-        if (posizionePrecedente == -1) {
-            posizionePrecedente = position;
-            quantitaIngredientePrecedente = quantitaIngrediente;
-            ingredientePrecedente = ingrediente;
-        }
-
-    }
-    public void controlloCambioSelezione(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-        if (posizionePrecedente != position) {
-            aggiornaListaIngrediente(verificaIngrediente(ingredientePrecedente, quantitaIngredientePrecedente), posizionePrecedente);
-            posizionePrecedente = position;
-            quantitaIngredientePrecedente = quantitaIngrediente;
-            ingredientePrecedente = ingrediente;
-        }
-
-    }
 
 
-    public void togliQuantitaIngrediente(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        if (ingrediente.getQuantitaPosseduta() < 10 && quantitaBottone(position) == 10) {
-            ingrediente.setQuantitaPosseduta(0);
-        } else {
-            ingrediente.setQuantitaPosseduta(verificaIngrediente(ingrediente ,quantitaIngrediente).getQuantitaPosseduta() - quantitaBottone(position));
-        }
-        quantitaIngrediente.setText(ingrediente.getQuantitaAssolutaToString());
 
-    }
-
-    public void aggiungiQuantitaIngrediente(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        ingrediente.setQuantitaPosseduta(verificaIngrediente(ingrediente, quantitaIngrediente).getQuantitaPosseduta() + quantitaBottone(position));
-        quantitaIngrediente.setText(ingrediente.getQuantitaAssolutaToString());
-    }
-
-    public void aggiornaListaIngrediente(Ingrediente ingrediente, int position){
-        listaIngredientiRicetta.get(position).setQuantitaPosseduta(ingrediente.getQuantitaPosseduta());
-    }
-
-    public void resetQuantitaLasciatoTestoVuoto(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-        if (quantitaIngrediente.getText().length() == 0) {
-            ingrediente.setQuantitaPosseduta(0);
-            aggiornaListaIngrediente(ingrediente, position);
-        }
-    }
-
-    public void verificaNumeroLitriBirra(EditText numeroLitriBirra, boolean hasFocus){
-        if(!hasFocus) {
-            if (numeroLitriBirra.getText().toString().isEmpty()) {
-                numeroLitriBirra.setText("0");
-            } else {
-                numeroLitriBirra.setText(String.valueOf(Integer.parseInt(numeroLitriBirra.getText().toString())));
-            }
-        }
-    }
-
-    public boolean controlloCreazione(View view,EditText nomeRicetta, EditText numeroLitriBirra ){
-        if(nomeRicetta.getText().toString().isEmpty()){
-            Snackbar.make(view, R.string.nome_ricetta_mancante, LENGTH_SHORT).show();
-            return false;
-        } else if( Double.parseDouble(numeroLitriBirra.getText().toString()) == 0.0 ){
-            Snackbar.make(view, R.string.litri_birra_ricetta_mancante, LENGTH_SHORT).show();
-            return false;
-        }
-        return true;
-    }
 
     public void salvaRicetta(View view, int zeroIngredinti ) {
         if (zeroIngredinti < 3) {
@@ -209,12 +123,5 @@ public class CreaRicettaFragment extends Fragment {
         }
     }
 
-    public void rispostaInvioTastiera(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        quantitaIngrediente.setOnKeyListener((v, keyCode, event) -> {
-            if (event.getAction() == KeyEvent.ACTION_DOWN && keyCode == KeyEvent.KEYCODE_ENTER) {
-                aggiornaListaIngrediente(verificaIngrediente(ingrediente, quantitaIngrediente), position);
-            }
-            return false;
-        });
-    }
+
 }
