+++ app/src/main/java/it/unimib/brewday/ui/gestisci_ricette/RicettaDettagliataFragment.java
@@ -8,7 +8,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -24,6 +23,7 @@ import it.unimib.brewday.R;
 import it.unimib.brewday.databinding.FragmentRicettaDettagliataBinding;
 import it.unimib.brewday.model.Ingrediente;
 import it.unimib.brewday.ui.gestisci_ingredienti.AdapterListViewListaIngredientiDisponibili;
+import it.unimib.brewday.util.GestioneRicette;
 import it.unimib.brewday.util.ListaIngredienti;
 
 
@@ -34,10 +34,8 @@ public class RicettaDettagliataFragment extends Fragment {
     private FragmentRicettaDettagliataBinding fragmentRicettaDettagliataBinding;
     boolean visibile;
 
-    int posizionePrecedente = -1;
-    EditText quantitaIngredientePrecedente;
+    GestioneRicette gestioneRicette;
 
-    Ingrediente ingredientePrecedente;
     public RicettaDettagliataFragment() {
         // Required empty public constructor
     }
@@ -67,6 +65,7 @@ public class RicettaDettagliataFragment extends Fragment {
         Button modificaRicettaButton = fragmentRicettaDettagliataBinding.buttonModificaRicetta;
         EditText numeroLitriRicettaBirra = fragmentRicettaDettagliataBinding.editTextNumberLitriRicettaBirra;
         EditText nomeRicetta = fragmentRicettaDettagliataBinding.textViewNomeRicetta;
+         gestioneRicette = new GestioneRicette();
 
 
         ListaIngredienti listaIngredienti = new ListaIngredienti();
@@ -83,20 +82,18 @@ public class RicettaDettagliataFragment extends Fragment {
         );
 
         numeroLitriRicettaBirra.setOnFocusChangeListener((v, hasFocus) ->
-                verificaNumeroLitriBirra(numeroLitriRicettaBirra, hasFocus)
+                gestioneRicette.verificaNumeroLitriBirra(numeroLitriRicettaBirra, hasFocus)
         );
 
 
 }
 
     public void setVisibile(boolean invertiVisibile, View view, EditText numeroLitriBirra, List<Ingrediente> listaIngredientiRicetta){
-
-
         if (!invertiVisibile){
             fragmentRicettaDettagliataBinding.buttonModificaRicetta.setText(R.string.conferma);
             visibile = !invertiVisibile;
         } else {
-            if( controlloCreazione(view, fragmentRicettaDettagliataBinding.textViewNomeRicetta, fragmentRicettaDettagliataBinding.editTextNumberLitriRicettaBirra)) {
+            if( gestioneRicette.controlloCreazione(view, fragmentRicettaDettagliataBinding.textViewNomeRicetta, fragmentRicettaDettagliataBinding.editTextNumberLitriRicettaBirra)) {
                 int zeroIngredinti = 0;
                 double litriScelti = Double.parseDouble(numeroLitriBirra.getText().toString());
                 List<Double> listaIngredientiPerLitro = new ArrayList<>();
@@ -141,96 +138,6 @@ public class RicettaDettagliataFragment extends Fragment {
         fragmentRicettaDettagliataBinding.listViewIngredrientiRicettaDettagliata.setDivider(null);
     }
 
-    public Ingrediente verificaIngrediente(Ingrediente ingrediente, EditText quantitaIngrediente) {
-
-        if (quantitaIngrediente.getText().length() == 0) {
-            ingrediente.setQuantitaPosseduta(0);
-            quantitaIngrediente.setText("0");
-        } else {
-            quantitaIngrediente.setText(String.valueOf(Integer.parseInt(quantitaIngrediente.getText().toString())));
-            ingrediente.setQuantitaPosseduta(Integer.parseInt(quantitaIngrediente.getText().toString()));
-        }
-
-        return ingrediente;
-    }
-
-    public int quantitaBottone(int position) {
-        if (position == 0) {
-            return 1;
-        } else {
-            return 10;
-        }
-
-
-    }
-    public void inizializzaPositionePrecedente(Ingrediente ingrediente, int position, EditText quantitaIngrediente) {
-
-        if (posizionePrecedente == -1) {
-            posizionePrecedente = position;
-            quantitaIngredientePrecedente = quantitaIngrediente;
-            ingredientePrecedente = ingrediente;
-        }
-
-    }
-    public void controlloCambioSelezione(Ingrediente ingrediente, int position, EditText quantitaIngrediente, List<Ingrediente> listaIngredientiRicetta) {
-
-        if (posizionePrecedente != position) {
-            aggiornaListaIngrediente(verificaIngrediente(ingredientePrecedente, quantitaIngredientePrecedente), posizionePrecedente, listaIngredientiRicetta);
-            posizionePrecedente = position;
-            quantitaIngredientePrecedente = quantitaIngrediente;
-            ingredientePrecedente = ingrediente;
-        }
-
-    }
-
-
-    public void togliQuantitaIngrediente(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        if (ingrediente.getQuantitaPosseduta() < 10 && quantitaBottone(position) == 10) {
-            ingrediente.setQuantitaPosseduta(0);
-        } else {
-            ingrediente.setQuantitaPosseduta(verificaIngrediente(ingrediente ,quantitaIngrediente).getQuantitaPosseduta() - quantitaBottone(position));
-        }
-        quantitaIngrediente.setText(ingrediente.getQuantitaAssolutaToString());
-
-    }
-
-    public void aggiungiQuantitaIngrediente(Ingrediente ingrediente, int position, EditText quantitaIngrediente){
-        ingrediente.setQuantitaPosseduta(verificaIngrediente(ingrediente, quantitaIngrediente).getQuantitaPosseduta() + quantitaBottone(position));
-        quantitaIngrediente.setText(ingrediente.getQuantitaAssolutaToString());
-    }
-
-    public void aggiornaListaIngrediente(Ingrediente ingrediente, int position, List<Ingrediente> listaIngredientiRicetta){
-        listaIngredientiRicetta.get(position).setQuantitaPosseduta(ingrediente.getQuantitaPosseduta());
-    }
-
-    public void resetQuantitaLasciatoTestoVuoto(Ingrediente ingrediente, int position, EditText quantitaIngrediente, List<Ingrediente> listaIngredientiRicetta) {
-        if (quantitaIngrediente.getText().length() == 0) {
-            ingrediente.setQuantitaPosseduta(0);
-            aggiornaListaIngrediente(ingrediente, position, listaIngredientiRicetta);
-        }
-    }
-
-    public void verificaNumeroLitriBirra(EditText numeroLitriBirra, boolean hasFocus){
-        if(!hasFocus) {
-            if (numeroLitriBirra.getText().toString().isEmpty()) {
-                numeroLitriBirra.setText("0");
-            } else {
-                numeroLitriBirra.setText(String.valueOf(Integer.parseInt(numeroLitriBirra.getText().toString())));
-            }
-        }
-    }
-
-    public boolean controlloCreazione(View view,EditText nomeRicetta, EditText numeroLitriBirra ){
-        if(nomeRicetta.getText().toString().isEmpty()){
-            Snackbar.make(view, R.string.nome_ricetta_mancante, LENGTH_SHORT).show();
-            return false;
-        } else if( Double.parseDouble(numeroLitriBirra.getText().toString()) == 0.0 ){
-            Snackbar.make(view, R.string.litri_birra_ricetta_mancante, LENGTH_SHORT).show();
-            return false;
-        }
-        return true;
-    }
-
     public void salvaRicetta(View view, int zeroIngredinti , List<Ingrediente> listaIngredientiRicetta, List<Double> listaIngredientiPerLitro) {
         if (zeroIngredinti < 3) {
             //TODO chiamata luca con listaIngredientiPerLitro
@@ -241,12 +148,5 @@ public class RicettaDettagliataFragment extends Fragment {
         }
     }
 
-    public void rispostaInvioTastiera(Ingrediente ingrediente, int position, EditText quantitaIngrediente, List<Ingrediente> listaIngredientiRicetta){
-        quantitaIngrediente.setOnKeyListener((v, keyCode, event) -> {
-            if (event.getAction() == KeyEvent.ACTION_DOWN && keyCode == KeyEvent.KEYCODE_ENTER) {
-                aggiornaListaIngrediente(verificaIngrediente(ingrediente, quantitaIngrediente), position, listaIngredientiRicetta);
-            }
-            return false;
-        });
-    }
+
 }
